+++ tests/psd_tools2/api/test_pil_io.py
@@ -35,4 +35,4 @@ def test_get_color_mode(mode):
     (ColorMode.LAB, False, 'LAB'),
 ])
 def test_get_pil_mode(mode, alpha, expected):
-    assert pil_io.get_pil_mode(mode, alpha) == expected
+    assert pil_io.get_pil_mode(mode.name, alpha) == expected